[IMP] chart: change colorscale definition type from string to colors
This commit changes the way we define the colorscale in the charts
definition. Instead of an object [min, mid, max] OR a string, each
colorscale is now defined by a 3-colors scheme. This allow a simpler
definition as well as the ability for the user to select a classical
colorscale and edit quickly one of the 3 colors.

Task: 4816317